requirements.txt
torch
requests
numpy
tqdm
scikit-image
lpips
pyiqa
opencv-python
thop
pytorch-wavelets
pytz
PyWavelets
pandas
pillow
scikit-learn
scipy
spikingjelly
setuptools